Refactor following the Stroustrup guidelines
    This is a huge commit and its main purpose was to align the source code
to Stroustrup guidelines, though I don't know if it's fully compliant to
it.
    Removed somme inner blocks of scope in some switch statements and
changed the way LLVM files are included in the project, instead of the
quote symbol I'm adopting the angle brackets because the LLVM are found
in the path and not in the project tree.
    Also were added documentation entries with some information in it to
documentation be later generated. But there is a lot to do, a real deep
look at the code is needed to improve it.

Signed-off-by: Rafael Campos Nunes <rafaelnunes@engineer.com>