Remove libm requirement from generated code
Also as nothing is used from it remove the unnecessary <math.h> header.

This replaces the runtime calls to the log10() function with a
precalculated constant.

It depends on the fact that you already have Log base 256 of the maximum
integer size (called sizeof) and simply converts that to a log10 by
multiplying by 2.41 (log(256)/log(10) --> 2.408239965311849) and adding
the necessary padding for sign, nul and the first digit.